rake task creates/installs coffeescript.syntax
The ultraviolet gem doesn't work with Ruby 1.9, instead
we need to use the updated 'spox' versions of ultraviolet
and plist:

  spox-ultraviolet, spox-plist

This new rake task: install_coffeescript_syntax loads
either the original gems if running Ruby < 1.9 or the
'spox' versions if running on Ruby >= 1.9.

Load CoffeeScript.tmLanguage directly from the github repo
for the TextMate bundle.

Parse and write the yaml-format syntax file directly
into the correct location in the ultraviolet gem.