Reduce brackets of If statement (Optimization)
Reduce brackets of If statement, which reduces the compiled js size.
Add a hasReturn variable to detect whether the block makes a return
expression. unwrap the block, and only reduce brackets when
subexpression is Assign, Call or Return.

We can't reduce brackets when subexpression is For statement,
Because For statement makes variable declaration for index variable.