Remove reference to never-defined .isUndefined property
This was added in f4a7cca0755b61fe70db0f7dc5b841aec44022ae to fix #1038
for CoffeeScript 1.0.1. `.isUndefined` was removed in
caf3da2f664f65f3aead70a6292385dea18da9ba but this code was never
updated. That actually caused the behavior of this code to change
(trailing `undefined` and `return undefined` statements no longer got
optimized away) when CoffeeScript 1.3.2 was released, but the new
behavior was deemed correct in
https://github.com/jashkenas/coffee-script/issues/1038#issuecomment-14427560.