Update example to treat strings as strings
One of the current examples for destructuring assignment, the futurist poet, invokes an alert statement using `alert("name + "-" + street");`. This causes **name** and **street** to be incorrectly cast to Number, resulting in a reader seeing "NaN".

Instead, the leading and trailing `"` should be removed, so that the `-` operator is not used, and instead `+` concatenates the strings correctly.

This fixes that change, changing the code to `alert(name + "-" + street);`.