Update critic example to treat strings as strings
One of the current examples for destructuring assignment, the literary critic, invokes an alert statement using `alert("first + " " + last");`. This causes the alert to not fire due to a SyntaxError.

Instead, the leading and trailing `"` should be removed, so that `+` concatenates the strings correctly.

This fixes that change, changing the code to `alert(first + " " + last);`. It also changes the text visible to the reader to match the amended code.