CI: Ensure browser tests are up-to-date
Today I learned about the cake rule `doc:test` which builds the browser tests from the main tests (similar to `build:browser` but for the tests). This PR ensures that those tests are up-to-date before CI can succeed, similar to the other build operations we do.

As a side note, I find the name `doc:test` rather confusing, as it's about building tests not documenting or running them. But that's less important.